under mem

+++ ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: akrid <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/01 15:48:02 by akrid             #+#    #+#             */
-/*   Updated: 2023/11/01 16:01:08 by akrid            ###   ########.fr       */
+/*   Updated: 2023/11/03 12:16:45 by akrid            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *s)
+#include "libft.h"
+
+size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
